Fix: what_time_is_it_now to use CLOCK_MONOTONIC when preset (macOS 10.12+) Fix: compile error for bound_image because of implicit conversion from int to float Fix: _snprintf is not defined when __APPLE__. Defined _snprintf as snprintf

+++ src/data.c
@@ -355,7 +355,7 @@ box bound_image(image im)
             }
         }
     }
-    box b = {minx, miny, maxx-minx + 1.0f, maxy-miny + 1.0f};
+    box b = {static_cast<float>(minx), static_cast<float>(miny), maxx-minx + 1.0f, maxy-miny + 1.0f};
     //printf("%f %f %f %f\n", b.x, b.y, b.w, b.h);
     return b;
 }
